+++ src/components/MemoriWidget/MemoriWidget.tsx
@@ -2968,6 +2968,9 @@ const MemoriWidget = ({
     showOnlyLastMessages === undefined
       ? selectedLayout !== 'TOTEM' && selectedLayout !== 'WEBSITE_ASSISTANT'
       : !showOnlyLastMessages;
+  const canShowLoginButton =
+    !tenant?.ssoLogin &&
+    (showLogin ?? integrationConfig?.showLogin ?? memori.requireLoginToken);
 
   const headerProps: HeaderProps = {
     memori: {
@@ -3001,8 +3004,7 @@ const MemoriWidget = ({
     showReload: selectedLayout === 'TOTEM',
     showClear: showClear ?? integrationConfig?.showClear ?? false,
     clearHistory: () => setHistory(h => h.slice(-1)),
-    showLogin:
-      showLogin ?? integrationConfig?.showLogin ?? memori.requireLoginToken,
+    showLogin: canShowLoginButton,
     setShowLoginDrawer,
     loginToken,
     user,
@@ -3062,7 +3064,7 @@ const MemoriWidget = ({
     isUserLoggedIn: !!loginToken && !!user?.userID,
     hasInitialSession: !!initialSessionID,
     notEnoughCredits: needsCredits && !hasEnoughCredits,
-    showLogin: showLogin ?? memori.requireLoginToken,
+    showLogin: canShowLoginButton,
     setShowLoginDrawer,
     showChatHistory:
       showChatHistory ?? integrationConfig?.showChatHistory ?? true,